Turn one of the temutil::get_timeseries(...) in to a template.
Struggled to get the explicit instantiation call setup right in the .cpp
file. ALthough slightly more complicated, it maket the header look a lot
cleaner, and also prevents having to include the Logging stuff in the
TEMUtilityFunctions header.

This should *almost* compile. The problem is that the C netcdf interface
requires specific function calls for specific types of variable. One idea
(noted in comments) is to use nc_inq_var() to figure out the type and then
call the right function based on type?