OK, this is quite ugly, but actually working!!!
But there are some problems. The biggest problem that I know of right
now is that the workers processes spew out a bunch of data (all the
OutputDataNuggets) and without the arbitrary sleep() command when a cell
is done running, there is not enought time for the master to recieve all
the outstanding requests before it gets the "cell done" message, and
ends up telling the workers to quit and everyone exits.

Not sure how to handle this...for curiosity, it would be interesting
to try MPI_SSend (synchronous send mode) where the call does not complete
before the reciever gets the message. But this is not supported by
boost::mpi, and we are using boost::mpi for the serialization.

One other idea might be to put something on the master side that tracks
how long it has been since the last request handle was recieved/processed.
Then we could assume that if the done_q is full and it has been something
like a minute since the last good request was recieved, that it would be
safe to assume that all the OutputDataNuggets have been received and it
would be safe to quit...