Make command line option for forcing outputs to full size (of time dim).
Previously, as an attempt to work better with parallel netcdf, we had
been creating output files with no unlimited dimension and a pre-defined
size such that the files would accomodate all the run-years specified for
the run.

Now the default behavior is to use time as the unlimited dimension, and
you can override that to pre-set the file size using the command line flag.