master/io worker/computer worker refactor. has HARDCODED values!
This is a fairly substantial refactor to the parallel MPI code to try a
different arrangement of process types.

In this attempt, I tried to use 3 process types:

 - Master
 - IO worker
 - Compute worker

Right now it is hardcoded to have 3 IO worker processes, each
responsible for writing to roughly a 3rd of the alaphebet (so IO
process 1 writes to files beginning with letters A-J, process 2 writes
 to files starting with K-T, etc.)

The other feature of this attempt is that I tried to track the time
that is spent *after* all the cells are complete, but *before* all the
data has been run thru the IO processors.

Apart from some tiny testing runs, I did one "full size" run of our
sample 10x10 area on atlas. I used the batch splitting/merging.
Oddly it looked like the middle IO processor (~K-T) took by far the
longest? I was able to look at this by looking at the slurm-NNNNNN.out
files and noting when (roughly) each process recieved a DIE command from
master (indicating that all the cells were completed). Then I can
watch the output directory, sorting by time to see roughtly what is
going on:

    $ watch ls -alh --sort time batch-run/batch-0/output/

It looks like in general the IO workers are way slower than the
compute nodes, so they compute nodes post all their messages and it
takes quite a while for the IO workers to get to reading all the
messages and writing to the files.

Unfortunately it took about 4:45hrs to complete! Less than one hour
was required to run all the cells. The remainder was waiting for
the IO workers (in each batch) to finish. This was run using the
batch script and using a batch size of 10 (with 14 processors per
batch).