Change cmd line option for nc-output-last-n-eq
Now general nc output (config file, 'output_nc_eq')  must be enabled for
this option to have effect.

Also add this option to the config file. The command line takes
precedence over the config file value. If not specifed on either command
line or config file, then the default is to output the entire eq stage
(if nc output is enabled generally).